configure: Support for HEASLR on mingw targets
The appropriate flag for HEASLR (--high-entropy-va) is scheduled for
inclusion in the next version of binutils (2.25), doesn't hurt to
include it a little early.

Also set the image base >4GB so higher entropy gets applied to image
base randomization when used with HEASLR (8 -> 17 bits of
randomization).  Windows does this for compatibility because of "latent
pointer truncation issues".

Signed-off-by: Alex Smith <alex.smith@warpsharp.info>